Third Person Camera Base Implement
Third person Camera using Cinemachine now has rotation around the player. The player now has 2 new transforms one further ahead of the player for the virtual camera to look at while still looking at the world The other transform is a orientation specifically for the camera to follow so that the rotation can be applied to the cam orientation without the rotation affecting the original player orientation. now the camera has dedicated buttons for the user to rotate both clock wise and counter clock wise and even reset back to the original camera position